requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.12.1
pymysql
cryptography
httpx==0.25.1
twilio==8.10.2
websockets
